requirements.txt
flask==1.1.2
werkzeug==1.0.1
tornado==4.5.3
PyYAML==5.4.1
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-Assets==2.0
pyserial==3.5
libusb1==1.9.2
netaddr==0.8.0
watchdog==0.10.6
sarge==0.1.6
requests==2.25.1
requests-toolbelt==0.9.1
flask-compress==1.9.0
netifaces==0.10.9
ws4py==0.5.1
futures==3.3.0
blinker==1.4
v4l2==0.2
pycrypto==2.6.1
pyudev==0.22.0
paho-mqtt==1.5.1
